cti: rework Object to use encoding stacks
Previous implementation allowed only for a single encoding step.
Type of encoder was stored in an Attr field, which was used to guide
the encoding process.

Eventhough it allowed for storing multiple encoder info's, it did not
preserve the order in which each encoding shall be perfomed.

New implementation drops Attr and Aux (unused) fields and introduces a
one - the Encoding field. The field is a stack of encoder types used to encode
given field in the object tree. Each time the same field is encoded in
the same expansion step, new encoding info is pushed to the stack.
Complementary - each time it is decoded, the info is pop from the stack.

The serialization (expansion/compaction) process stays still an all-or-nothing
operation - meaning a field can only be fully encoded or fully decoded.
If expansion fails in the middle of a serialization, it may leave an
object tree in a dirty state. This shortcomming can be worked around by
copying the object prior to serialization.

This allows for stacking multiple encoders for even a single node. In
practice the new Object is now capable of representing not only json/yaml/etc
files, but also archives that contain them.